% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_funs.R
\name{clean_comp}
\alias{clean_comp}
\title{Clean comparison matrix}
\usage{
clean_comp(comp)
}
\arguments{
\item{comp}{Dataframe with comparisons}
}
\value{
A cleaned dataframe
}
\description{
\code{clean_comp()} prepares the comparison file so that all names
match names in the metadata, and clean the names so that there are no
conflicts.
}
