% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_funs.R
\name{clean_key}
\alias{clean_key}
\title{Clean metadata}
\usage{
clean_key(metadata)
}
\arguments{
\item{metadata}{Dataframe with metadata}
}
\value{
A cleaned dataframe
}
\description{
\code{clean_key()} prepares the metadata so that all names match
names in the data matrix, and clean the names so that there are no
conflicts.
}
