% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_funs.R
\name{cluster_plot}
\alias{cluster_plot}
\title{Plots cluster profile.}
\usage{
cluster_plot(df, meta, hc, k)
}
\arguments{
\item{df}{Input dataframe. Raw data matrix, with protein ids and sample
abundances.}

\item{meta}{Input dataframe. Metadata associated to \code{df}.}

\item{hc}{An object of class "hclust" created with the \code{dend_calc()}
function.}

\item{k}{Number of clusters.}
}
\value{
A ggplot2 object.
}
\description{
\code{cluster_plot()} plots a profile of the specified cluster of phosphopeptides. Values a z-score transformed previous to plotting.
}
