% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funs.R
\name{cofounders_plot}
\alias{cofounders_plot}
\title{Principal component analysis (or PCA) regression plot}
\usage{
cofounders_plot(res.pca)
}
\arguments{
\item{res.pca}{Output list from \code{pca_prot()}}
}
\value{
A ggplot2 object
}
\description{
\code{cofounders_plot()} uses the information provided by calculating
a PCA, extracts the top principal components (PCs), which are orthogonal
variables explaining the information contained within it. Then, the
different clinical or technical variables indicated in the metadata are
regressed against the PCs. The heatmap shows the values of the R² and
significance (FDR), indicating the strength of the relationship between the
variables and the PCs.
}
