% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pn_funs.R
\name{create_hyp_obj}
\alias{create_hyp_obj}
\title{Perform Fisher's exact test}
\usage{
create_hyp_obj(bio_list, enrich_list, df, method_p_val)
}
\arguments{
\item{bio_list}{List populated by vectors containing the gene symbols of
interest}

\item{enrich_list}{PN database as a list, results from
\code{create_pn_enrich_lists()}}

\item{df}{Original dataframe from where \code{bio_list} comes from. It serves as
the background for creating the enrichment list}

\item{method_p_val}{Method to calculate de P value from the hypergeometric
distribution, either using the minimum-likelihood approach ('minlike',
default), or doubling approach (mid-P value, 'central')}
}
\value{
A list of lists, with dataframes for each category (branch or class)
and any other extra clustering (e.g., up or down), with the following
variables: \code{label} (specific branch or class), \code{pval} (P value),
\code{signature} (number of DE genes), \code{geneset} (number of genes in the
geneset), \code{overlap} (number of DE genes found in the geneset), \code{background}
(number of genes in the experiment), \code{hits} (readable version of overlap).
}
\description{
\code{create_hyp_obj()} performs Fisher's exact test (using either
the minimum-likelihood or doubling approach) and creates result dataframe
}
