% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_funs.R
\name{dend_calc}
\alias{dend_calc}
\title{Performs hierarchical clustering and plots number of clusters}
\usage{
dend_calc(df, meta, plot = TRUE)
}
\arguments{
\item{df}{Input dataframe. Raw data matrix, with protein ids and sample
abundances.}

\item{meta}{Input dataframe. Metadata associated to \code{df}.}

\item{plot}{Boolean indicating if silhouette plot should be displayed or not.
Default is TRUE.}
}
\value{
An object of class "hclust".
}
\description{
\code{dend_calc()} performs hierarchical clustering and prints a
silhouette plot to help choose number of clusters.
}
