% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_funs.R
\name{dotplot_enrich}
\alias{dotplot_enrich}
\title{Creates dotplot for GO enrichment results}
\usage{
dotplot_enrich(ego)
}
\arguments{
\item{ego}{An enrichResult instance, output from \code{enrich_prot()} function.}
}
\value{
A ggplot2 object.
}
\description{
Creates dotplot for GO enrichment results
}
