% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pn_funs.R
\name{enrich_pn_plot}
\alias{enrich_pn_plot}
\title{Enrichment dot plot}
\usage{
enrich_pn_plot(hyp_df, method_p_adjust)
}
\arguments{
\item{hyp_df}{List resulting from \code{prep_hyp_df()}}

\item{method_p_adjust}{Type of adjusted P value to use in the dot plot}
}
\value{
A ggplot2 figure, a dot plot, showing the branch/class labels, the
fold enrichment, adjusted P value and faceted (or not) depending on the
cluster/grouping.
}
\description{
\code{enrich_pn_plot()} creates a dot plot showcasing the PN
enrichment
}
