% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_fun.R
\name{export_results}
\alias{export_results}
\title{Export limma result(s)}
\usage{
export_results(limma_res, df_pre_limma, metadata, phospho = "no", org = NULL)
}
\arguments{
\item{limma_res}{Input list. Output from the function \code{limma_de()}.}

\item{df_pre_limma}{Input dataframe. Data that is fed to \code{limma_de()}, should
be normalized and, optionally, imputed.}

\item{metadata}{Input dataframe. Metadata associated to the data.}

\item{phospho}{Input vector of length 1. Either 'yes' or 'no' (default).}

\item{org}{Input vector of length 1. Either 'human' or 'mouse'. Default NULL.}
}
\value{
A dataframe containing annotation, log2-abundances of each sample,
logFC, p value, and adjusted p value for each comparison
}
\description{
\code{export_results()} is a function that prepares the results from
differential expression - performed with \code{{limma}} - to be exported to excel.
It creates a single dataframe with annotation, log2-expression
}
