% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{filt_na}
\alias{filt_na}
\title{Filters dataframe based on missing values}
\usage{
filt_na(df, threshold = 0.7)
}
\arguments{
\item{df}{dataframe to be filtered}

\item{threshold}{numeric input indicating the minimum percentage of observed values across all samples. Goes from 0 to 1, default is 0.7 (i.e., 70\%)}
}
\value{
a filtered dataframe
}
\description{
Filters dataframe based on missing values
}
