% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funs.R
\name{heatmap_prot}
\alias{heatmap_prot}
\title{Plots heatmap with color-coded annotation.}
\usage{
heatmap_prot(df, meta, vars, z.score = FALSE, cutrow = NA, cutcol = NA)
}
\arguments{
\item{df}{Input dataframe. Data matrix.}

\item{meta}{Input dataframe. Metadata associated to \code{df}.}

\item{vars}{Input character vector of length >= 1. Indicates the variables
that are going to be}

\item{z.score}{Input boolean vector. Default is FALSE. Performs z score.}

\item{cutrow}{Input numeric vector. Indicates the k groups or clusters that the row dendrogram should be divided into.}

\item{cutcol}{Input numeric vector. Indicates the k groups or clusters that the column dendrogram should be divided into.}
}
\value{
A pheatmap object.
}
\description{
\code{heatmap_prot()} plots a heatmap with extra annotation in the
form of colors. Wrapper for \code{pheatmap()}.
}
