% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_funs.R
\name{import}
\alias{import}
\title{Import TMT proteomic data}
\usage{
import(
  input_file_name,
  metadata = NULL,
  type = NULL,
  org = NULL,
  tmt = NULL,
  level = NULL
)
}
\arguments{
\item{input_file_name}{String vector with name(s) of the files, including
their extension}

\item{metadata}{Dataframe with associated metadata}

\item{type}{Either "psm" or "protein", indicating the level of the input data
\itemize{
\item only relevant if the input is several files, if only one file, then NULL
}}

\item{org}{Organism - only relevant if the input is several files, if only
one file, then NULL}

\item{tmt}{Either "tmt", "tmtpro" or "tmtpro0", indicating}

\item{level}{Either "peptide" or "protein", indicating the desired
output-level data - only relevant if the input is several files, if only
one file, then NULL}
}
\value{
A dataframe at the complexity level specified by the user
}
\description{
\code{import()} imports one or several files of TMT proteomic data,
from several formats (txt, csv, tsv). If several files are indicated, and
depending on the complexity level (peptide spectrum matches (PSM) or
protein) it can collapse the data to the desired level (either peptide or
protein). It's also TMT-sensitive (TMT, TMT-pro or TMT-0) if input data is
at the PSM-level.
}
