% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_funs.R
\name{irs_norm}
\alias{irs_norm}
\title{Internal reference scaling (IRS) normalization}
\usage{
irs_norm(df, meta, tmm = TRUE)
}
\arguments{
\item{df}{Dataframe that is going to be IRS-normalized}

\item{meta}{Dataframe of associated metadata. Pool channels should be flagged
as 'Norm', but if a single sample is used repeatedly across all TMTs as a
sort of technical replicate that acts as a reference channel, and these
samples should be then collapsed into a single one after normalization,
they should be flagged as 'Ref'.}

\item{tmm}{TRUE (default) or FALSE, whether the function should perform TMM
normalization after SL normalization and before IRS normalization}
}
\value{
A normalized dataset
}
\description{
\code{irs_norm()} performs internal reference scaling (IRS)
normalization (after sample loading (SL) and TMM (optional) normalizations)
in the presence of 2 or more TMT experiments. It uses one (or several)
reference/pool channels, and in the absence of one, calculates a 'virtual'
one using the information of all proteins of the plexes.
}
