% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinswing_funs.R
\name{kinswing_datable}
\alias{kinswing_datable}
\title{Summarized results from KinSwing analysis}
\usage{
kinswing_datable(ks_list)
}
\arguments{
\item{ks_list}{Input list. Output list of \code{kinswing()} function.}
}
\value{
A list of dataframe(s) with the predicted phosphopeptide-kinase
relationships per comparison.
}
\description{
\code{kinswing_datable()} generates a list with the results of the
KinSwing analysis, that is, it reports the predicted phosphopeptide-kinase
relationships.
}
