% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinswing_funs.R
\name{kinswing_plot}
\alias{kinswing_plot}
\title{Volcano plot of predicted kinases}
\usage{
kinswing_plot(ks_list, coefs)
}
\arguments{
\item{ks_list}{Input list. Output list of \code{kinswing()} function.}

\item{coefs}{Input vector. A character vector created with \code{transform_comp()}
from ProteoMarker.}
}
\value{
A ggplot2 object.
}
\description{
\code{kinswing_plot()} returns a ggplot2 object, a volcano plot of
the output of the \code{kinswing()} function.
}
