% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_wrapper.R
\name{limma_de}
\alias{limma_de}
\title{limma wrapper function}
\usage{
limma_de(df.norm, meta, comp, org, phospho = "no", covars = NULL, anova)
}
\arguments{
\item{df.norm}{input dataframe or matrix. Ideally it is normalized, and it
will be subject to log2 transformation}

\item{meta}{dataframe with the metadata}

\item{comp}{dataframe with the comparisons of interest. Must have a Condition
and Control variables}

\item{org}{organism. Either "human", "mouse" or "yeast"}

\item{phospho}{is it a phosphoproteomics analysis or not (default "no")}

\item{covars}{input character vector indicating blocking/confounding
variable(s), default is NULL}

\item{anova}{input logical to indicate whether topTable should output results
from each comparison or as an ANOVA}
}
\value{
A list of dataframes, one for each comparison (or ANOVA) with the
results yielded by topTable
}
\description{
\code{limma_de()} is a wrapper function of limma, and performs the
whole pipeline.
}
