% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funs.R
\name{pca_prot}
\alias{pca_prot}
\title{Computes PCA}
\usage{
pca_prot(df, meta, scale = FALSE)
}
\arguments{
\item{df}{input dataframe}

\item{meta}{dataframe with the metadata}

\item{scale}{logical. Whether scale the data or not (default FALSE)}
}
\value{
a list containing the results from computing the PCA, as well as any
confounder variables and the transformed dataframe.
}
\description{
\code{pca_prot()} calculates principal component analysis (PCA) using
stats::prcomp(). This function removes any missing values and then prepares
the input dataframe so that prcomp() can use it.
}
