% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funs.R
\name{plot_pca}
\alias{plot_pca}
\title{Plots PCA}
\usage{
plot_pca(res.pca, var_p)
}
\arguments{
\item{res.pca}{list as output from pca_prot()}

\item{var_p}{string. Variable from metadata to color.}
}
\value{
a ggplot2 object (list)
}
\description{
Plots PCA
}
