% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pn_funs.R
\name{prep_hyp_df}
\alias{prep_hyp_df}
\title{Modify the enrichment dataframe}
\usage{
prep_hyp_df(hyp_obj, level)
}
\arguments{
\item{hyp_obj}{list resulting from \code{create_hyp_obj()}}

\item{level}{either 'class' or 'branch'}
}
\value{
A dataframe as the one from \code{fisher_enrich()}, but with the
following extra variables: \code{fdr} (adjusted P value by BH method),
\code{bonferroni} (adjusted P value by Bonferroni method), \code{cluster} (extra
grouping, if any), \code{BgRatio} (number of all genes in specific term / number
of universal genes), \code{GeneRatio} (number of genes enriched in specific term
/ number of input genes), \code{FoldEnrichment} (GeneRatio / BgRatio)
}
\description{
\code{prep_hyp_df()} takes the resulting dataframe from the
enrichment analysis, calculates the adjusted P value, as well as the
BgRatio, GeneRatio and the FoldEnrichment. It also makes the class labels
readable
}
