% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_funs.R
\name{process_raw}
\alias{process_raw}
\title{Clean proteomic data matrix}
\usage{
process_raw(software, raw, metadata, phospho = FALSE)
}
\arguments{
\item{software}{String of length 1 indicating the software of origin, either
FragPipe ("fp"), MaxQuant ("mq") or others ("other").}

\item{raw}{Dataframe with the proteomic data. If "other" was selected in
\code{software}, the matrix should contain only the numeric matrix where the
column names correspond to what's indicated in the metadata; and a column
with the protein identifications (called either "Index", "Accession" or
"Protein.IDs").}

\item{metadata}{Dataframe with the corresponding metadata.}

\item{phospho}{TRUE or FALSE (default), whether \code{raw} is a phosphoproteomic
data matrix.}
}
\value{
A dataframe with the proteomic data ready to work with.
}
\description{
\code{process_raw()} prepares the proteomic data matrix, depending on
its origin (FragPipe, MaxQuant, or others), so that it can be work with
safely. It cleans variable names, transforms 0 into NA, and removes
contaminants.
}
