% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funs.R
\name{prot_barplot}
\alias{prot_barplot}
\title{Barplot for proteomics data}
\usage{
prot_barplot(df, metadata, fill = "Condition")
}
\arguments{
\item{df}{Dataframe to be plotted}

\item{metadata}{Dataframe with associated metadata}

\item{fill}{String with variable name to be used as filling/grouping variable}
}
\value{
A ggplot2 object
}
\description{
\code{prot_barplot()} creates barplots of proteomics data
}
