% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funs.R
\name{prot_boxplot}
\alias{prot_boxplot}
\title{Boxplots for proteomics data}
\usage{
prot_boxplot(df, metadata, fill, reorder = FALSE, var_reorder = "counts")
}
\arguments{
\item{df}{Dataframe to be plotted}

\item{metadata}{Dataframe with associated metadata}

\item{fill}{String with name of variable to fill the boxplots}

\item{reorder}{TRUE or FALSE (default). Reorder boxplots based on intensity}

\item{var_reorder}{String with name of the variable to use for reorder}
}
\value{
A ggplot2 object
}
\description{
\code{prot_boxplot()} creates boxplots for all samples - can be
reordered or not, based on the intensity/abundance variable
}
