% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funs.R
\name{prot_denplot}
\alias{prot_denplot}
\title{Density plot for proteomics data}
\usage{
prot_denplot(df, metadata, color = "BioReplicate")
}
\arguments{
\item{df}{Dataframe to be plotted}

\item{metadata}{Dataframe with associated metadata}

\item{color}{String with variable name to be used as grouping variable}
}
\value{
A ggplot2 object
}
\description{
\code{prot_denplot()} creates density plots of proteomics data
}
