% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{prot_log2}
\alias{prot_log2}
\title{Log2-transformation}
\usage{
prot_log2(df, meta)
}
\arguments{
\item{df}{dataframe to be log2-transformed}

\item{meta}{dataframe with metadata}
}
\value{
a log2-transformed dataframe
}
\description{
Log2-transformation
}
