% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psp}
\alias{psp}
\title{PhosphoSitePlus kinase-substrate mouse database}
\format{
A data frame with 2264 rows and 2 variables:
\describe{
\item{kinase}{Kinase gene symbols}
\item{substrate}{Aminoacid sequence (14aa) of the corresponding substrate}
}
}
\source{
\url{https://www.phosphosite.org/homeAction.action}
}
\usage{
psp
}
\description{
A dataset containing a paired list of kinases and their respective
substrates. The variables are as follows:
}
\keyword{datasets}
