% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_wrapper.R
\name{result_table}
\alias{result_table}
\title{Formatting of results from limma}
\usage{
result_table(res)
}
\arguments{
\item{res}{input list resulting from limma_de()}
}
\value{
A dataframe
}
\description{
\code{result_table()} generates a dataframe with a summary of the
number of significantly changing proteins, both upregulated and
downregulated, along with the number of those that do not change. The
cutoff is based on the adjusted P value (< 0.05)
}
