% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_funs.R
\name{select_cluster}
\alias{select_cluster}
\title{Assignment of phosphopeptides to respective clusters.}
\usage{
select_cluster(hc, k = NULL, df)
}
\arguments{
\item{hc}{An object of class "hclust" created with the \code{dend_calc()}
function.}

\item{k}{Number of clusters.}

\item{df}{Input dataframe. Raw data matrix, with protein ids and sample
abundances.}
}
\value{
A list of dataframes with as many levels as specified in \code{k}.
}
\description{
\code{select_cluster()} assigns each phosphopeptide to its cluster,
based on the number of clusters selected by the user.
}
