% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_funs.R
\name{SL_norm}
\alias{SL_norm}
\title{Sample loading normalization}
\usage{
SL_norm(df, meta)
}
\arguments{
\item{df}{dataframe with the raw data (not log2-transformed)}

\item{meta}{dataframe with metadata}
}
\value{
a normalized dataframe
}
\description{
\code{SL_norm()} normalizes the sum of each channel to the average
grand total
}
