% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_funs.R
\name{tmm_norm}
\alias{tmm_norm}
\title{calcNormFactors wrapper}
\usage{
tmm_norm(df, meta)
}
\arguments{
\item{df}{dataframe with the raw data (not log2-transformed), usually as
output of SL_norm()}

\item{meta}{dataframe with metadata}
}
\value{
a normalized dataframe
}
\description{
\code{tmm_norm()} is a wrapper function for edgeR::calcNormFactors().
It accounts for compositional bias in the data.
}
