% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmt_integrator.R
\name{tmt_integrator}
\alias{tmt_integrator}
\title{FragPipe's TMT Integrator wrapper}
\usage{
tmt_integrator(
  psm,
  metadata,
  org,
  tmt,
  use.unique = FALSE,
  purity_thr = 0.5,
  pep_prob_thr = 0.9,
  sum_ms2_int_thr = 0.05
)
}
\arguments{
\item{psm}{Input FragPipe data matrix at the PSM level.}

\item{metadata}{Input dataframe. Metadata associated to \code{psm}.}

\item{org}{Input vector of length >= 1. Indicates the species (could be several) the data belongs
to.}

\item{tmt}{Input vector of length 1. Indicates the type of TMT used, either
\code{tmt}, \code{tmtpro} or \code{tmtpro0}.}

\item{use.unique}{Whether to use only unique peptides for quantification. Default is FALSE.}

\item{purity_thr}{Numeric vector of length 1. Indicates purity threshold.
Default is 0.5 (50\%).}

\item{pep_prob_thr}{Numeric vector of length 1. Indicates peptide probability
threshold. Default is 0.9.}

\item{sum_ms2_int_thr}{Numeric vector of length 1. Indicates the percentage
of MS2 intensity below which PSMs are going to be filtered out. Default is
0.05 (5\%). For PTM data, such as phosphorylation, it should be 0.025
(2.5\%).}
}
\value{
A list containing both peptide and protein level information
}
\description{
\code{tmt_integrator()} mimics the FragPipe's TMT Integrator module
filters for TMT data, minus de PSM normalization. This includes purity and
peptide probability filters, as well as a minimum summed MS2 intensity,
presence of TMT and, optionally, PTM labeling. It summarises PSM
information to the peptide and subsequent protein level, for further
analysis.
}
