% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funs.R
\name{volcano_prot}
\alias{volcano_prot}
\title{Plots volcano plot}
\usage{
volcano_prot(df, var, anno)
}
\arguments{
\item{df}{Input dataframe. Output dataframe from \code{limma_de()}.}

\item{var}{Input character vector of lenght 1. Either \code{P.Value} or
\code{adj.P.Val}.}

\item{anno}{Input character vector of lenght equal to the number of rows of
\code{df}. Determines the type of labels (uniprot id, symbol...) marking the
points of the volcano plot.}
}
\value{
A ggplot2 object.
}
\description{
\code{volcano_prot()} creates a volcano plot using the output
dataframe from \code{limma_de()}.
}
