% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_funs.R
\name{wgcna_prot}
\alias{wgcna_prot}
\title{WGCNA wrapper function}
\usage{
wgcna_prot(
  df,
  meta,
  Group,
  traits,
  var = NULL,
  org,
  RCutoff = 0.85,
  MCutheight = 0.1,
  PowerUpper = 30,
  minModuleSize = 20,
  verbose = TRUE
)
}
\arguments{
\item{df}{Input dataframe. Data matrix with normalized (log2) abundances.}

\item{meta}{Input dataframe. Metadata associated to \code{df}.}

\item{Group}{Input dataframe. Dataframe with the grouping, should have two
columns: Sample and Group.}

\item{traits}{Input dataframe. For module-trait correlation, should have
Sample column, plus the rest of clinical variables, dummy-encoded (0-1),
and controls are not included.}

\item{var}{Input character vector of length 1, optional. Extra variable to
take into account. Default is NULL.}

\item{org}{Input character vector of length 1. Either "human" or "mouse".}

\item{RCutoff}{Cut-off value for the correlation R square, used to adjust the
soft threshold selection, default is 0.85.}

\item{MCutheight}{Height in the dendrogram, establishes if modules are going
to be merged, default is 0.1.}

\item{PowerUpper}{Maximum soft-thresholding power to choose from, default is
30.}

\item{minModuleSize}{Controls the minimum number of proteins in each cluster
\itemize{
\item if a few hundred/a thousand proteins, should be between 20-30, more than
that, 50. Default is 20.
}}

\item{verbose}{Print intermediate steps. Default is TRUE.}
}
\value{
A list of lists, with figures and the proteins that conform each
cluster.
}
\description{
\code{wgcna_prot()} is a wrapper function of several functions from
the well-known \code{{WGCNA}} package, that are run in a pipeline to conduct the
WGCNA analysis, start to end. It outputs the figures and list with clusters.
}
